Add events --follow --update, fix deploy.
There was an issue dealing with errors in the concurrent code in the
deploy command (i.e. events + signalling). The concurrency issue is now
addressed by a combination of futures and promises and the "wait"
functionality is now included in the events command.

Wait functionality is to be added to the signal command in a subsequent
commit.